refactor: dto패키지 분리, dto에 serialName추가
