+++ android/app/src/main/java/com/zzang/chongdae/data/mapper/GroupPurchaseDetailResponseMapper.kt
@@ -1,9 +1,8 @@
 package com.zzang.chongdae.data.mapper
 
-import com.zzang.chongdae.data.remote.dto.GroupPurchaseDetailResponse
+import com.zzang.chongdae.data.remote.dto.response.GroupPurchaseDetailResponse
 import com.zzang.chongdae.domain.model.ArticleDetail
 import com.zzang.chongdae.domain.model.CurrentCount
-import com.zzang.chongdae.domain.model.Participation
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
