+++ android/app/src/main/java/com/zzang/chongdae/data/mapper/ParticipationsResponseMapper.kt
@@ -1,6 +1,6 @@
 package com.zzang.chongdae.data.mapper
 
-import com.zzang.chongdae.data.remote.dto.ParticipationResponse
+import com.zzang.chongdae.data.remote.dto.response.ParticipationResponse
 import com.zzang.chongdae.domain.model.Participation
 
 fun ParticipationResponse.toDomain() = Participation(
