--- android/app/src/main/java/com/zzang/chongdae/data/remote/dto/GroupPurchasesResponse.kt
@@ -1,9 +1,0 @@
-package com.zzang.chongdae.data.remote.dto
-
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class GroupPurchasesResponse(
-    val responses: List<GroupPurchase>,
-)
