--- android/app/src/main/java/com/zzang/chongdae/data/remote/dto/ParticipationResponse.kt
@@ -1,9 +1,0 @@
-package com.zzang.chongdae.data.remote.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ParticipationResponse(
-    val currentCount: Int,
-    val status: PurchaseStatus,
-)
