+++ android/app/src/main/java/com/zzang/chongdae/data/remote/dto/request/ParticipationRequest.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.data.remote.dto
+package com.zzang.chongdae.data.remote.dto.request
 
 import kotlinx.serialization.Serializable
