+++ android/app/src/main/java/com/zzang/chongdae/data/remote/dto/response/GroupPurchasesResponse.kt
@@ -1,0 +1,10 @@
+package com.zzang.chongdae.data.remote.dto.response
+
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GroupPurchasesResponse(
+    @SerialName("responses") val responses: List<GroupPurchase>,
+)
