+++ android/app/src/main/java/com/zzang/chongdae/data/remote/dto/response/PurchaseStatus.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.data.remote.dto
+package com.zzang.chongdae.data.remote.dto.response
 
 import kotlinx.serialization.SerialName
