+++ android/app/src/main/java/com/zzang/chongdae/data/repository/remote/GroupPurchaseRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.zzang.chongdae.data.repository.remote
 
 import com.zzang.chongdae.data.mapper.toDomain
-import com.zzang.chongdae.data.remote.dto.ParticipationRequest
+import com.zzang.chongdae.data.remote.dto.request.ParticipationRequest
 import com.zzang.chongdae.data.remote.source.GroupPurchaseDataSource
 import com.zzang.chongdae.domain.model.Article
 import com.zzang.chongdae.domain.model.ArticleDetail
